Add Symdef ChildList

+++ src/main/java/de/verschwiegener/mvr/GeneralSceneDescription.java
@@ -56,9 +56,9 @@ public class GeneralSceneDescription {
     @XmlElement(name = "Scene", required = true)
     protected Scene scene;
     @XmlAttribute(name = "verMajor", required = true)
-    protected BigInteger verMajor;
+    protected int verMajor;
     @XmlAttribute(name = "verMinor", required = true)
-    protected BigInteger verMinor;
+    protected int verMinor;
     @XmlAttribute(name = "provider", required = true)
     protected String provider;
     @XmlAttribute(name = "providerVersion", required = true)
@@ -120,7 +120,7 @@ public class GeneralSceneDescription {
      *     {@link BigInteger }
      *     
      */
-    public BigInteger getVerMajor() {
+    public int getVerMajor() {
         return verMajor;
     }
 
@@ -132,7 +132,7 @@ public class GeneralSceneDescription {
      *     {@link BigInteger }
      *     
      */
-    public void setVerMajor(BigInteger value) {
+    public void setVerMajor(int value) {
         this.verMajor = value;
     }
 
@@ -144,7 +144,7 @@ public class GeneralSceneDescription {
      *     {@link BigInteger }
      *     
      */
-    public BigInteger getVerMinor() {
+    public int getVerMinor() {
         return verMinor;
     }
 
@@ -156,7 +156,7 @@ public class GeneralSceneDescription {
      *     {@link BigInteger }
      *     
      */
-    public void setVerMinor(BigInteger value) {
+    public void setVerMinor(int value) {
         this.verMinor = value;
     }
